requirements.txt
django~=1.11
MySQL-python~=1.2.5
djangorestframework~=3.6.3
drf-nested-routers~=0.90.0